// packages/sdk/src/generated/index-vault/types/assetEntry.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getStructDecoder,
  getStructEncoder,
  getU16Decoder,
  getU16Encoder,
  getU64Decoder,
  getU64Encoder,
  getU8Decoder,
  getU8Encoder,
  type Address,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";
import {
  getAssetKindDecoder,
  getAssetKindEncoder,
  type AssetKind,
  type AssetKindArgs,
} from ".";

export type AssetEntry = {
  mint: Address;
  tokenProgram: Address;
  oracle: Address;
  targetWeightBps: number;
  kind: AssetKind;
  decimals: number;
  /** Internal bookkeeping; direct transfers to the vault ATA are ignored. */
  balance: bigint;
};

export type AssetEntryArgs = {
  mint: Address;
  tokenProgram: Address;
  oracle: Address;
  targetWeightBps: number;
  kind: AssetKindArgs;
  decimals: number;
  /** Internal bookkeeping; direct transfers to the vault ATA are ignored. */
  balance: number | bigint;
};

export function getAssetEntryEncoder(): FixedSizeEncoder<AssetEntryArgs> {
  return getStructEncoder([
    ["mint", getAddressEncoder()],
    ["tokenProgram", getAddressEncoder()],
    ["oracle", getAddressEncoder()],
    ["targetWeightBps", getU16Encoder()],
    ["kind", getAssetKindEncoder()],
    ["decimals", getU8Encoder()],
    ["balance", getU64Encoder()],
  ]);
}

export function getAssetEntryDecoder(): FixedSizeDecoder<AssetEntry> {
  return getStructDecoder([
    ["mint", getAddressDecoder()],
    ["tokenProgram", getAddressDecoder()],
    ["oracle", getAddressDecoder()],
    ["targetWeightBps", getU16Decoder()],
    ["kind", getAssetKindDecoder()],
    ["decimals", getU8Decoder()],
    ["balance", getU64Decoder()],
  ]);
}

export function getAssetEntryCodec(): FixedSizeCodec<
  AssetEntryArgs,
  AssetEntry
> {
  return combineCodec(getAssetEntryEncoder(), getAssetEntryDecoder());
}